Portability fixes for Solaris and OSF1.
git-svn-id: svn://svn.code.sf.net/p/rdesktop/code/trunk/rdesktop@11 423420c4-83ab-492f-b58f-81f9feb106b5

+++ orders.c
@@ -842,7 +842,7 @@ void process_orders(STREAM s)
 	}
 
 	if (s->p != next_packet)
-		WARN("%d bytes remaining\n", next_packet - s->p);
+		WARN("%d bytes remaining\n", (int)(next_packet - s->p));
 }
 
 /* Reset order state */
